added seistools/legacy
